requirements.txt
streamlit
supabase
Pillow
pandas
pyyaml
webdav4
fsspec